requirements.txt
streamlit==1.45.1
opencv-python==4.8.0.76
librosa==0.10.0
scenedetect==0.6.2
ffmpeg-python==0.2.0
python-dotenv==1.0.0
moviepy==1.0.3
pydub==0.25.1
torch==2.0.0
torchvision==0.15.0
torchaudio==2.0.0
yt_dlp==2024.4.9
google-api-python-client==2.113.0
